refactor: rename data-collector module to container-data-collector and localhost-data-collector to machine-data-collector
