+++ metrics-backend/container-data-collector/.dockerignore
