+++ metrics-backend/container-data-collector/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+spring.application.name=container-data-collector
+
+server.port = 8001
+spring.kafka.bootstrap-servers=${BOOTSTRAP_SERVER}
