+++ metrics-backend/container-data-collector/src/test/java/kr/cs/interdata/containerdatacollector/ContainerResourceMonitorTest.java
@@ -1,0 +1,23 @@
+package kr.cs.interdata.containerdatacollector;
+
+import org.junit.jupiter.api.Test;
+
+public class ContainerResourceMonitorTest {
+
+    @Test
+    public void testCollectContainerResources() {
+        // 메서드 호출
+        //String json = ContainerResourceMonitor.collectContainerResources();
+
+        // JSON 파싱
+        //Gson gson = new Gson();
+        //JsonObject jsonObject = gson.fromJson(json, JsonObject.class);
+
+        // 필수 키 값 존재 여부 확인
+       // assertNotNull(jsonObject.get("containerId"), "containerId는 null이 아니어야 합니다.");
+      //  assertTrue(jsonObject.get("cpuUsagePercent").isJsonPrimitive(), "cpuUsagePercent는 숫자여야 합니다.");
+       // assertTrue(jsonObject.get("memoryTotalBytes").isJsonPrimitive(), "memoryTotalBytes는 숫자여야 합니다.");
+       // assertTrue(jsonObject.get("diskReadBytes").isJsonPrimitive(), "diskReadBytes는 숫자여야 합니다.");
+      //  assertNotNull(jsonObject.get("network"), "network 정보는 null이 아니어야 합니다.");
+    }
+}
