+++ metrics-backend/container-data-collector/src/test/java/kr/cs/interdata/containerdatacollector/DataCollectorApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.cs.interdata.containerdatacollector;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DataCollectorApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
