--- metrics-backend/data-collector/build.gradle
@@ -1,90 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.4.4'
-    id 'io.spring.dependency-management' version '1.1.7'
-    id 'application'
-}
-
-bootJar {
-    enabled = true
-}
-jar {
-    enabled = false
-}
-
-group = 'kr.cs.interdata'
-version = '0.0.1-SNAPSHOT'
-
-
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-springBoot {
-    //mainClass = 'kr.cs.interdata.datacollector.ResourceMonitorDaemon'
-    mainClass = 'kr.cs.interdata.datacollector.DataCollectorApplication'
-}
-
-application {
-    mainClass = 'kr.cs.interdata.datacollector.DataCollectorApplication'
-    //mainClass = 'kr.cs.interdata.datacollector.ResourceMonitorDaemon'
-}
-
-tasks.register("runLocalHostMonitor", JavaExec) {
-    group = "application"
-    description = "Run ResourceMonitorDaemon main"
-    classpath = sourceSets.main.runtimeClasspath
-    mainClass = 'kr.cs.interdata.datacollector.ResourceMonitorDaemon'
-}
-
-tasks.register("runContainerMonitor", JavaExec) {
-    group = "application"
-    description = "Run DataCollectorApplication main"
-    classpath = sourceSets.main.runtimeClasspath
-    mainClass = 'kr.cs.interdata.datacollector.DataCollectorApplication'
-}
-
-
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(':metrics-backend:producer')
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.kafka:spring-kafka'
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.kafka:spring-kafka-test'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-
-    compileOnly 'org.projectlombok:lombok:1.18.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.30'
-
-    implementation 'com.github.oshi:oshi-core:6.4.5'
-    implementation 'com.google.code.gson:gson:2.10.1'
-    implementation "com.h2database:h2"
-
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.2'
-
-    implementation 'com.github.docker-java:docker-java:3.5.2'
-    implementation 'com.github.docker-java:docker-java-core:3.5.2'
-    implementation 'com.github.docker-java:docker-java-transport-httpclient5:3.5.2'
-
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
