--- metrics-backend/data-collector/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-spring.application.name=data-collector
-
-server.port = 8001
-spring.kafka.bootstrap-servers=${BOOTSTRAP_SERVER}
