--- metrics-backend/data-collector/src/test/java/kr/cs/interdata/datacollector/DataCollectorApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.cs.interdata.datacollector;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DataCollectorApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
