--- metrics-backend/localhost-data-collector/build.gradle
@@ -1,52 +1,0 @@
-plugins {
-    id 'java'
-    id 'application'
-}
-
-group = 'kr.cs.interdata'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-application {
-    mainClass = 'kr.cs.interdata.localhostdatacollector.LocalhostDataCollectorApplication'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // 기존 의존성
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.kafka:spring-kafka'
-    compileOnly 'org.projectlombok:lombok'
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.kafka:spring-kafka-test'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
-    // 테스트용 JUnit 5 (선택, Spring Boot Starter Test에 이미 JUnit 5가 포함됨)
-    // testImplementation platform('org.junit:junit-bom:5.10.0')
-    // testImplementation 'org.junit.jupiter:junit-jupiter'
-
-    // 시스템 정보 수집용 OSHI
-    implementation 'com.github.oshi:oshi-core:6.4.5'
-
-    // JSON 직렬화용 Gson
-    implementation 'com.google.code.gson:gson:2.10.1'
-
-    // JSON 예쁘게 출력용 Jackson
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.2'
-
-}
-
-
-test {
-    useJUnitPlatform()
-}
