+++ metrics-backend/machine-data-collector/.dockerignore
